re #2708: simplify filtering of types, get rid of an unlikely edge case
The first implementation catered to the event that an existing docpool
was modified and an app added that would bring a new type, which would
then have to be integrated into the immediately addable types without
overriding any possible manual configuration. Now, we decided to assume that
such a complicated thing won't happen.