re #4825: remove menu code not currently reachable
docpool.menu.browser.menu used to contain a number of configuration-like
variables which could, however, not be manipulated other than directly
in that code. This wasn't done in 7 years; maybe there had once been
another way to override those values at runtime but if so and if the
functionality is ever going to be desired again, we can always bring the
deleted code back from git history and will then likely have to rewrite
it anyway.