re #5821: don't reverse order of groups in content area menu
Simply removing the reverse() seems to restore the sorting order of that
drop-down menu to what it was in Plone 5.1, to the extent possible to compare
using the screenshots attached to the ticket. It is not at all clear to me why
I had introduced that reversal anyway, the commit which did it (012a4386)
really doesn't explain it.